requirements.txt
cloudpickle==2.2.1
graphviz==0.20.3
numpy==2.2.1
pandas==2.2.2
py4j==0.10.9.7
pyarrow==17.0.0
pyspark==3.5.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.17.0
tzdata==2024.2
